Move the test into testdata
