+++ src/NUnitFramework/testdata/TestContextOneTimeTearDownData.cs
@@ -1,0 +1,37 @@
+// Copyright (c) Charlie Poole, Rob Prouse and Contributors. MIT License - see LICENSE.txt
+
+using NUnit.Framework;
+using NUnit.Framework.Interfaces;
+
+namespace NUnit.TestData
+{
+    [TestFixture]
+    public class TestContextOneTimeTearDownData
+    {
+        public TestContext.ResultAdapter ResultInOneTimeTearDown { get; private set; }
+
+        [Test]
+        public void TestTruth()
+        {
+            Assert.That(true, Is.True);
+        }
+
+        [Test]
+        public void TestFalsehood()
+        {
+            Assert.That(false, Is.False);
+        }
+
+        [Test, Explicit]
+        public void TestExplicit()
+        {
+            Assert.Pass("Always passes if you run it!");
+        }
+
+        [OneTimeTearDown]
+        public void OneTimeTearDown()
+        {
+            ResultInOneTimeTearDown = TestContext.CurrentContext.Result;
+        }
+    }
+}
